+++ modules/vehicle/__init__.py
@@ -1,4 +1,5 @@
 from modules.vehicle.manager import VehicleManager
 from modules.vehicle.models import VehicleStatus
+from modules.vehicle.state_engine import merge_vehicle_state
 
-__all__ = ["VehicleManager", "VehicleStatus"]
+__all__ = ["VehicleManager", "VehicleStatus", "merge_vehicle_state"]